# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Emergency Dispatch System Backend API
  version: 1.0.0
  description: |
    Comprehensive OpenAPI specification for all backend services in the
    Emergency Dispatch System.

    Services covered:
    - Auth Service
    - Incident Service
    - Dispatch Service
    - Analytics Service
    - Realtime Gateway (WebSocket handshake endpoint)
servers:
  - url: http://localhost:8081
    description: Auth service (local)
  - url: http://localhost:8082
    description: Incident service (local)
  - url: http://localhost:8083
    description: Dispatch service (local)
  - url: http://localhost:8084
    description: Analytics service (local)
  - url: http://localhost:8085
    description: Realtime gateway (local)
...
security:
  - bearerAuth: []
paths:
...
  /vehicles/register:
    post:
      tags: [Dispatch]
      summary: Register vehicle
      operationId: registerVehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterVehicleRequest'
      responses:
        '201':
          description: Vehicle created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /vehicles:
    get:
      tags: [Dispatch]
      summary: List vehicles
      operationId: listVehicles
      responses:
        '200':
          description: Vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vehicle'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /vehicles/available:
    get:
      tags: [Dispatch]
      summary: List available vehicles
      operationId: listAvailableVehicles
      parameters:
        - name: station_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Available vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vehicle'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /vehicles/nearest:
    get:
      tags: [Dispatch]
      summary: Internal nearest available vehicle lookup
      security: []
      operationId: getNearestAvailableVehicle
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: lng
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [ambulance, fire_truck, police_car]
      responses:
        '200':
          description: Nearest vehicle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /vehicles/{id}:
    parameters:
      - $ref: '#/components/parameters/UUIDPathId'
    get:
      tags: [Dispatch]
      summary: Get vehicle by ID
      operationId: getVehicle
      responses:
        '200':
          description: Vehicle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Dispatch]
      summary: Update vehicle metadata
      operationId: updateVehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVehicleRequest'
      responses:
        '200':
          description: Vehicle updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Dispatch]
      summary: Delete vehicle
      operationId: deleteVehicle
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /vehicles/{id}/location:
    parameters:
      - $ref: '#/components/parameters/UUIDPathId'
    get:
      tags: [Dispatch]
      summary: Get current vehicle location
      operationId: getVehicleLocation
      responses:
        '200':
          description: Current location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleLocationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Dispatch]
      summary: Update vehicle location
      operationId: updateVehicleLocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLocationRequest'
      responses:
        '200':
          description: Location updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleLocationUpdateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /vehicles/{id}/status:
    parameters:
      - $ref: '#/components/parameters/UUIDPathId'
    put:
      tags: [Dispatch]
      summary: Update vehicle status
      operationId: updateVehicleStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVehicleStatusRequest'
      responses:
        '200':
          description: Vehicle updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UUIDPathId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    RegisterVehicleRequest:
      type: object
      properties:
        station_id:
          type: string
          format: uuid
        station_type:
          type: string
          enum: [hospital, police, fire]
        vehicle_type:
          type: string
          enum: [ambulance, fire_truck, police_car]
        license_plate:
          type: string
        driver_name:
          type: string
        driver_id:
          type: string
          format: uuid
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
      required: [station_id, station_type, vehicle_type]
    Vehicle:
      type: object
      properties:
        id:
          type: string
          format: uuid
        station_id:
          type: string
          format: uuid
        station_type:
          type: string
          enum: [hospital, police, fire]
        vehicle_type:
          type: string
          enum: [ambulance, fire_truck, police_car]
        license_plate:
          type: string
        driver_name:
          type: string
        driver_id:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum: [available, en_route, at_scene, returning, off_duty]
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        incident_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UpdateVehicleRequest:
      type: object
      properties:
        station_id:
          type: string
          format: uuid
        station_type:
          type: string
          enum: [hospital, police, fire]
        vehicle_type:
          type: string
          enum: [ambulance, fire_truck, police_car]
        license_plate:
          type: string
        driver_name:
          type: string
        driver_id:
          type: string
          format: uuid
    UpdateLocationRequest:
      type: object
      properties:
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
      required: [latitude, longitude]
    UpdateVehicleStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [available, en_route, at_scene, returning, off_duty]
      required: [status]
    VehicleLocationResponse:
      type: object
      properties:
        vehicle_id:
          type: string
          format: uuid
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        status:
          type: string
        updated_at:
          type: string
          format: date-time
    VehicleLocationUpdateResponse:
      type: object
      properties:
        vehicle_id:
          type: string
          format: uuid
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        status:
          type: string
...